indentando o codigo e separando os styles
