--- src/context/OptionContext.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const OptionContext = React.createContext()
-
-export default OptionContext;
