+++ src/context/QuantityContext.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+const QuantityContext = React.createContext()
+
+export default QuantityContext;
