+++ src/screens/AddressPage/styles.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import {TextField} from "@material-ui/core";
+import { TextField } from "@material-ui/core";
 
 export const FormContainer = styled.form`
   height: 100%;
