+++ src/screens/CartPage/CartPage.js
@@ -14,21 +14,22 @@ import {
   ButtonConfirm,
   ButtonContinue,
   DivFormPayment
-} from "./styled";
+} from "./styles";
 import { Radio, RadioGroup, FormControlLabel } from "@material-ui/core";
 import { NavBottom } from "../../components/NavBottom/NavBottom"
 import NavBar from "../../components/NavBar/NavBar";
 import GlobalStateContext from "../../global/GlobalStateContext";
-import FoodCard from "../../components/FoodCard";
+import FoodCard from "../../components/FoodCard/FoodCard";
 import RestaurantContext from "../../context/RestaurantContext";
 import { useRequestData } from "../../hooks/useRequestData";
 import { BASE_URL } from "../../constants/urls";
 import { placeOrder } from "../../services/restaurant";
 import { useHistory } from "react-router-dom";
-import {goBack} from "../../router/coordinator"
+import { goBack } from "../../router/coordinator"
+
 const CartPage = () => {
   const { cart, setCart, subtotal, updateTotal } = useContext(GlobalStateContext)
-  const {restaurant} = useContext(RestaurantContext)
+  const { restaurant } = useContext(RestaurantContext)
   const getAddress = useRequestData(`${BASE_URL}/profile/address`, undefined)
   const [paymentMethod, setPaymentMethod] = useState("money")
   const history = useHistory()
@@ -54,7 +55,7 @@ const CartPage = () => {
 
   const setOrder = () => {
     const products = cart.map((product) => {
-      return {id: product.id, quantity: product.amount}
+      return { id: product.id, quantity: product.amount }
     })
 
     const body = {
@@ -70,39 +71,39 @@ const CartPage = () => {
       <NavBar />
       <DivAdress>
         <DeliveryAddress>Endereço de entrega</DeliveryAddress>
-        {getAddress &&  <Street>{getAddress.address.street} , {getAddress.address.number}</Street>}
+        {getAddress && <Street>{getAddress.address.street} , {getAddress.address.number}</Street>}
       </DivAdress>
       <DivCart>
         {cart.length === 0 ? <div>
           <Text>Carrinho Vazio</Text>
         </div> : <div>
-        <h1>{restaurant.name}</h1>
-        <h2>{restaurant.address}</h2>
-        <h3>{restaurant.deliveryTime} min</h3>
-        {cart.map((item) => {
-          return <DivOrder>
-          <FoodCard
-          key={item.id}
-          removeItemFromCart={() => removeItemFromCart(item)}
-          category={item.category}
-          description={item.description}
-          id={item.id}
-          name={item.name}
-          image={item.photoUrl}
-          price={item.price}
-          />
-        </DivOrder>
-        })}
-        </div>}
+            <h1>{restaurant.name}</h1>
+            <h2>{restaurant.address}</h2>
+            <h3>{restaurant.deliveryTime} min</h3>
+            {cart.map((item) => {
+              return <DivOrder>
+                <FoodCard
+                  key={item.id}
+                  removeItemFromCart={() => removeItemFromCart(item)}
+                  category={item.category}
+                  description={item.description}
+                  id={item.id}
+                  name={item.name}
+                  image={item.photoUrl}
+                  price={item.price}
+                />
+              </DivOrder>
+            })}
+          </div>}
       </DivCart>
 
-      <Freight>Frete R${subtotal!==0 ? `${restaurant.shipping}.00` : "00.00"}</Freight>
+      <Freight>Frete R${subtotal !== 0 ? `${restaurant.shipping}.00` : "00.00"}</Freight>
       <ButtonContinue onClick={() => goBack(history)}>
         Continuar comprando
       </ButtonContinue>
       <DivTotal>
         <Subtotal>SUBTOTAL</Subtotal>
-        <Price>R${subtotal!==0 ? (subtotal + restaurant.shipping) : "00.00"}</Price>
+        <Price>R${subtotal !== 0 ? (subtotal + restaurant.shipping) : "00.00"}</Price>
       </DivTotal>
       <DivFormPayment>
         <p>Forma de Pagamento</p>
@@ -115,7 +116,7 @@ const CartPage = () => {
       <ButtonConfirm onClick={setOrder}>
         Confirmar
       </ButtonConfirm>
-      <NavBottom changeColorCart={true}/>
+      <NavBottom changeColorCart={true} />
     </CartContainer>
   );
 };
