+++ src/screens/ProfilePage/ProfilePage.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { IconButton, Typography } from '@material-ui/core'
-import { BaseContainer, FlexBox, Divisor, ProgressStyled, ButtonLogout, ButtonBox, ProfilePageContainer, Text } from './profile-styles'
+import { IconButton } from '@material-ui/core'
+import { BaseContainer, FlexBox, Divisor, ProgressStyled, ButtonLogout, ButtonBox, ProfilePageContainer, Text } from './styles'
 import OrdesHistoryCard from '../../components/OrdersHistoryCard/OrdersHistoryCard'
 import EditIcon from '@material-ui/icons/Edit';
 import { useHistory } from 'react-router-dom';
@@ -13,62 +13,62 @@ import SettingsPowerIcon from '@material-ui/icons/SettingsPower';
 
 const ProfilePage = () => {
   const history = useHistory()
-  const userData = useRequestData(`${BASE_URL}/profile`, undefined) 
+  const userData = useRequestData(`${BASE_URL}/profile`, undefined)
   const ordersHistory = useRequestData(`${BASE_URL}/orders/history`, undefined)
 
   return (
     <div>
-       {userData ?
-    <ProfilePageContainer>
-      <NavBar />
-      <FlexBox>
-        <BaseContainer>
-          <Text variant="h6">{userData.user.name}</Text>
-          <IconButton onClick={() => goToEditProfile(history)}>
-            <EditIcon/>
-          </IconButton>
-        </BaseContainer>
-        <Text variant="h6">{userData.user.email}</Text>
-        <Text variant="h6">{userData.user.cpf}</Text>
-      </FlexBox>
-      <FlexBox greyBackground>
-        <BaseContainer>
-          <Text color="textSecondary" variant="h6">Endereço cadastrado</Text>
-          <IconButton onClick={() => goToAdress(history)}>
-            <EditIcon/>
-          </IconButton>
-        </BaseContainer>
-        <Text variant="h6">{userData.user.address}</Text>
-      </FlexBox>
-      <FlexBox>
-        <Text variant="h6">Histórico de pedidos</Text>
-        <Divisor/>
-      </FlexBox>
-      {ordersHistory && ordersHistory.orders.map((order, id) => {
-        console.log(order)
-        return (
-          <OrdesHistoryCard
-            key={id}
-            restaurantName={order.restaurantName}
-            date={order.createdAt}
-            price={order.totalPrice}
-          />
-        ) 
-      })}
-    <ButtonLogout onClick={() => logout(history)}>
-      <ButtonBox>
-        <SettingsPowerIcon />
+      {userData ?
+        <ProfilePageContainer>
+          <NavBar />
+          <FlexBox>
+            <BaseContainer>
+              <Text variant="h6">{userData.user.name}</Text>
+              <IconButton onClick={() => goToEditProfile(history)}>
+                <EditIcon />
+              </IconButton>
+            </BaseContainer>
+            <Text variant="h6">{userData.user.email}</Text>
+            <Text variant="h6">{userData.user.cpf}</Text>
+          </FlexBox>
+          <FlexBox greyBackground>
+            <BaseContainer>
+              <Text color="textSecondary" variant="h6">Endereço cadastrado</Text>
+              <IconButton onClick={() => goToAdress(history)}>
+                <EditIcon />
+              </IconButton>
+            </BaseContainer>
+            <Text variant="h6">{userData.user.address}</Text>
+          </FlexBox>
+          <FlexBox>
+            <Text variant="h6">Histórico de pedidos</Text>
+            <Divisor />
+          </FlexBox>
+          {ordersHistory && ordersHistory.orders.map((order, id) => {
+            console.log(order)
+            return (
+              <OrdesHistoryCard
+                key={id}
+                restaurantName={order.restaurantName}
+                date={order.createdAt}
+                price={order.totalPrice}
+              />
+            )
+          })}
+          <ButtonLogout onClick={() => logout(history)}>
+            <ButtonBox>
+              <SettingsPowerIcon />
         Sair
       </ButtonBox>
-    </ButtonLogout>
-    <NavBottom changeColorAvatar={true}/>
-    </ProfilePageContainer> :
-    <div>
-      <NavBar />
-      <ProgressStyled color="secondary"/>
-      <NavBottom changeColorAvatar={true}/>
-    </div>
-  }
+          </ButtonLogout>
+          <NavBottom changeColorAvatar={true} />
+        </ProfilePageContainer> :
+        <div>
+          <NavBar />
+          <ProgressStyled color="secondary" />
+          <NavBottom changeColorAvatar={true} />
+        </div>
+      }
     </div>
   )
 }
