+++ src/screens/ProfilePage/styles.js
@@ -8,7 +8,7 @@ export const FlexBox = styled.div`
     gap: 4px;
     flex-direction: column;
     align-items: flex-start;
-    background-color:${props => props.greyBackground ? "#EEEEEE" : "#FFF" };
+    background-color:${props => props.greyBackground ? "#EEEEEE" : "#FFF"};
 `
 
 export const BaseContainer = styled.div`
