+++ src/screens/RestaurantPage/styles.js
@@ -1,12 +1,12 @@
 import styled from 'styled-components'
 
-export const RestaurantContainer=styled.div`
+export const RestaurantContainer = styled.div`
   display:flex;
   flex-direction:column;
   margin-bottom: 72px;
-`  
+`
 
-export const RestaurantAll=styled.div`
+export const RestaurantAll = styled.div`
   display:flex;
   flex-direction: column;
   align-items: center;
